begin pipification of NucDiff
