+++ nucdiff/__version__.py
@@ -1,0 +1,3 @@
+VERSION = (0, 1, 0)
+
+__version__ = '.'.join(map(str, VERSION))
