+++ nucdiff/general.py
